requirements.txt
lorem-text
bottle
python-dotenv
SQLAlchemy
